docker-requirements.txt
imutils==0.5.4
keras==2.15.0
lazy_loader==0.3
matplotlib==3.8.3
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
pandas==2.2.1
PyAudio==0.2.14
pytesseract==0.3.10
seaborn==0.13.2
six==1.16.0
librosa==0.10.1
sounddevice==0.4.6
tensorboard==2.15.2
tensorboard-data-server==0.7.2
tensorflow-datasets==4.9.4
tensorflow-estimator==2.15.0
tensorflow-hub==0.16.1
tensorflow-io==0.36.0
tensorflow-io-gcs-filesystem==0.36.0
tf-keras==2.15.0
